pyproject.toml
[tool.poetry]
name = "reliable-rag-graph"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.scripts]
start = "reliable_rag_graph.server:start"

[tool.poetry.dependencies]
python = "^3.12"
langserve = {extras = ["all"], version = "^0.2.2"}
langchain = "^0.2.12"
langchain-core = "^0.2.28"
langchain-community = "^0.2.11"
langgraph = "^0.1.19"
langchain-openai = "^0.1.20"
python-dotenv = "^1.0.1"
python-multipart = "^0.0.9"
filetype = "^1.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
